Feat: update all controller actions to authenticate with api_key parameter against API_KEY environment variable; all tests passing

+++ app/controllers/api/v1/golfers_controller.rb
@@ -1,22 +1,28 @@
 class Api::V1::GolfersController < ApplicationController
   def index
-    golfers = Golfer.all
-    render json: GolferSerializer.new(golfers), status: 200
+    if params[:api_key] == ENV["API_KEY"]
+      golfers = Golfer.all
+      render json: GolferSerializer.new(golfers), status: 200
+    end
   end
 
   def create
-    golfer = Golfer.new(golfer_params)
-    if golfer.save
-      render json: GolferSerializer.new(golfer), status: :created
-    else
-      render json: {error: golfer.errors.full_messages.to_sentence}, status: 400
+    if params[:api_key] == ENV["API_KEY"]
+      golfer = Golfer.new(golfer_params)
+      if golfer.save
+        render json: GolferSerializer.new(golfer), status: :created
+      else
+        render json: {error: golfer.errors.full_messages.to_sentence}, status: 400
+      end
     end
   end
 
   def destroy
-    golfer = Golfer.find(params[:id])
-    golfer.destroy
-    render body: nil, status: :no_content
+    if params[:api_key] == ENV["API_KEY"]
+      golfer = Golfer.find(params[:id])
+      golfer.destroy
+      render body: nil, status: :no_content
+    end
   end
 
   private
